+++ R/optim-adagrad.R
@@ -36,20 +36,21 @@ optim_Adagrad <- R6::R6Class(
           param$state[['sum']]  <- torch_full_like(
             param, 
             initial_accumulator_value, 
-            memory_format=torch.preserve_format()
+            memory_format=torch_preserve_format()
           )
         }
       }
     },
     
-    share_memory = function(){
-      for (group in self$param_groups){
-        for (p in seq_along(group$params)) {
-          param <- group$params[[p]]
-          param$state[['sum']]$share_memory_()
-        }
-      }
-    },
+    # It's implemeneted in PyTorch, but it's not necessary at the moment
+    # share_memory = function(){
+    #   for (group in self$param_groups){
+    #     for (p in seq_along(group$params)) {
+    #       param <- group$params[[p]]
+    #       param$state[['sum']]$share_memory_()
+    #     }
+    #   }
+    # },
     
     step = function(closure = NULL){
       with_no_grad({
@@ -75,30 +76,33 @@ optim_Adagrad <- R6::R6Class(
             state_sum  <- param$state[['sum']]
             state_step <- param$state[['step']]
             
-           if (weight_decay != 0) {
+           if (group$weight_decay  != 0) {
               # if (grad$is_sparse) {
               #   runtime_error("weight_decay option is not compatible with sparse gradients")
               # }
-              grad <- grad$add(param, alpha = weight_decay)
+              grad <- grad$add(param, alpha = group$weight_decay)
             }
             
-            clr <-  lr / (1 + (step - 1) * lr_decay)
+            clr <-  group$lr / (1 + (param$state[['step']] - 1) * group$lr_decay)
             
-            if (grad$is_sparse) {
-              grad <- grad$coalesce()
-              grad_indices <- grad$`_indices`()
-              grad_values <-  grad$`_values`()
-              size <- grad$size()
+            # Sparse tensors handling will be added in future
+            # if (grad$is_sparse) {
+            #   grad <- grad$coalesce()
+            #   grad_indices <- grad$`_indices`()
+            #   grad_values <-  grad$`_values`()
+            #   size <- grad$size()
               
-              state_sum$add_(_make_sparse(grad, grad_indices, grad_values.pow(2)))
-              std = state_sum.sparse_mask(grad)
-              std_values = std._values().sqrt_().add_(eps)
-              param.add_(_make_sparse(grad, grad_indices, grad_values / std_values), alpha=-clr)
-            } else {
-              state_sum.addcmul_(grad, grad, value=1)
-              std = state_sum.sqrt().add_(eps)
-              param.addcdiv_(grad, std, value=-clr)
-            }
+              # state_sum$add_(`_make_sparse`(grad, grad_indices, grad_values.pow(2)))
+              # std <- param$state[['sum']]$sparse_mask(grad)
+              # std_values <- std$`_values()`$sqrt_()$add_(group$eps)
+              # param$add_(_make_sparse(grad, grad_indices, grad_values / std_values), alpha=-clr)
+            #} else {
+            
+            param$state[['sum']]$addcmul_(grad, grad, value = 1)
+            std <- param$state[['sum']]$sqrt()$add_(group$eps)
+            param$addcdiv_(grad, std, value =-clr)
+            
+            #}
             
           }
         }
@@ -119,16 +123,22 @@ optim_Adagrad <- R6::R6Class(
 #' @param eps (float, optional): term added to the denominator to improve
 #' numerical stability (default: 1e-10)
 #' 
+#' Adagrad is an especially good optimizer for sparse data.
+#' It individually modifies learning rate for every single parameter,
+#' dividing the original learning rate value by sum of the squares of the gradients.
+#' It causes that the rarely occurring features get greater learning rates.
+#' The main downside of this method is the fact that learning rate may be
+#' getting small too fast, so that at some point a model cannot learn anymore.
+#' 
 #' @note 
-#' 
 #' Update rule: 
-#' 
 #' deqn{
 #'\theta_{t+1} = \theta_{t} - \frac{\eta }{\sqrt{G_{t} + \epsilon}} \odot g_{t} 
 #' }
-#' 
-#' 
-#' 
+#' The equation above and some remarks quoted 
+#' after [*An overview of gradient descent optimization algorithms*](https://ruder.io/optimizing-gradient-descent/index.html#adagrad)
+#' by Sebastian Ruder.
+#' @export
 optim_adagrad <- function(params, lr=1e-2, lr_decay=0, weight_decay=0, 
                           initial_accumulator_value=0, eps=1e-10){
   optim_Adagrad$new(params, lr, lr_decay, weight_decay,